OVT-7411 set end date for old version of code
When new version has been accepted.  The date will be set to either a)
the date new version starts b) the date old version starts if it is in
future c) current date.

Had to change validation of endDate for both KoodiVersio and
KoodistoVersio for this. Rejects endDate now only if it is before
startDate.